# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Pametne Kuće API",
        "description": "API dokumentacija za upravljanje uređajima, korisnicima i ulogama (Admin, Običan, Dete).",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
...
        "/api/sobe": {
            "get": {
                "tags": [
                    "Sobe"
                ],
                "summary": "Izlistaj sve sobe",
                "operationId": "4fd56ae0be9a2b33f980a1f79bc35d64",
                "responses": {
                    "200": {
                        "description": "Uspešno učitane sobe sa podacima o stanu"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sobe"
                ],
                "summary": "Kreiraj novu sobu",
                "operationId": "37e7312ee496f8c85febd2ef6ecc79e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nazivSobe": {
                                        "type": "string",
                                        "example": "Dnevna soba"
                                    },
                                    "stan_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Soba uspešno kreirana"
                    },
                    "403": {
                        "description": "Zabranjeno (Uloga 'dete' ili niste vlasnik stana)"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sobe/{id}": {
            "get": {
                "tags": [
                    "Sobe"
                ],
                "summary": "Prikaži detalje sobe i uređaje",
                "operationId": "235a56609171d96d734901fdca81f5c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prikaz sobe sa stanjima svih uređaja"
                    },
                    "404": {
                        "description": "Soba nije pronađena"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sobe"
                ],
                "summary": "Ažuriraj podatke o sobi",
                "operationId": "94b938d9b8b309331cdbc0753efaa09e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nazivSobe": {
                                        "type": "string",
                                        "example": "Spavaća soba"
                                    },
                                    "stan_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Soba uspešno ažurirana"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sobe"
                ],
                "summary": "Obriši sobu",
                "operationId": "c58b1f0514c62a7938ee1137a8a0fbad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Soba obrisana"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ovde unesite access_token dobijen nakon logovanja.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}